# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartBus AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BusRoute {
  id: string;
  name: string;
  number: string;
  from: string;
  to: string;
  color: string;
  totalStops: number;
}

export interface BusStop {
  id: string;
  name: string;
  lat: number;
  lng: number;
  routeIds: string[];
}

export interface BusRouteDetail {
  id: string;
  name: string;
  number: string;
  from: string;
  to: string;
  color: string;
  stops: BusStop[];
  lastBusTime?: string | null;
  isLastBus: boolean;
}

export type LiveBusCrowdLevel =
  (typeof LiveBusCrowdLevel)[keyof typeof LiveBusCrowdLevel];

export const LiveBusCrowdLevel = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface LiveBus {
  id: string;
  routeId: string;
  routeNumber: string;
  routeName: string;
  routeColor: string;
  lat: number;
  lng: number;
  speed: number;
  heading: number;
  nextStop: string;
  nextStopId: string;
  distanceToNextStop: number;
  crowdLevel: LiveBusCrowdLevel;
  isLastBus: boolean;
  lastUpdated: string;
}

export type BusEtaCrowdLevel =
  (typeof BusEtaCrowdLevel)[keyof typeof BusEtaCrowdLevel];

export const BusEtaCrowdLevel = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface BusEta {
  busId: string;
  routeId: string;
  routeNumber: string;
  routeName: string;
  routeColor: string;
  etaMinutes: number;
  crowdLevel: BusEtaCrowdLevel;
  isLastBus: boolean;
}

export type CrowdPredictionLevel =
  (typeof CrowdPredictionLevel)[keyof typeof CrowdPredictionLevel];

export const CrowdPredictionLevel = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface CrowdPrediction {
  stopId: string;
  routeId?: string;
  level: CrowdPredictionLevel;
  estimatedPassengers: number;
  reason: string;
}

export interface FrequencyData {
  routeId: string;
  dayType: string;
  /** Buses per hour in morning (6-10am) */
  morning: number;
  /** Buses per hour in afternoon (10am-5pm) */
  afternoon: number;
  /** Buses per hour in evening (5-9pm) */
  evening: number;
  /** Buses per hour at night (9pm-midnight) */
  night: number;
}

export type SearchResultCrowdLevel =
  (typeof SearchResultCrowdLevel)[keyof typeof SearchResultCrowdLevel];

export const SearchResultCrowdLevel = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface SearchResult {
  routeId: string;
  routeNumber: string;
  routeName: string;
  routeColor: string;
  sourceStop: string;
  destinationStop: string;
  etaMinutes: number;
  crowdLevel: SearchResultCrowdLevel;
  isLastBus: boolean;
  frequency: number;
}

export type GetRouteFrequencyParams = {
  dayType?: GetRouteFrequencyDayType;
};

export type GetRouteFrequencyDayType =
  (typeof GetRouteFrequencyDayType)[keyof typeof GetRouteFrequencyDayType];

export const GetRouteFrequencyDayType = {
  weekday: "weekday",
  weekend: "weekend",
} as const;

export type GetStopCrowdParams = {
  routeId?: string;
};

export type SearchRoutesParams = {
  source: string;
  destination: string;
};