# lib/api-zod/src/generated/types/busEta.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartBus AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusEtaCrowdLevel } from "./busEtaCrowdLevel";

export interface BusEta {
  busId: string;
  routeId: string;
  routeNumber: string;
  routeName: string;
  routeColor: string;
  etaMinutes: number;
  crowdLevel: BusEtaCrowdLevel;
  isLastBus: boolean;
}