# lib/api-zod/src/generated/types/busRouteDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartBus AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusStop } from "./busStop";

export interface BusRouteDetail {
  id: string;
  name: string;
  number: string;
  from: string;
  to: string;
  color: string;
  stops: BusStop[];
  lastBusTime?: string | null;
  isLastBus: boolean;
}