# lib/api-zod/src/generated/types/crowdPrediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartBus AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CrowdPredictionLevel } from "./crowdPredictionLevel";

export interface CrowdPrediction {
  stopId: string;
  routeId?: string;
  level: CrowdPredictionLevel;
  estimatedPassengers: number;
  reason: string;
}