# lib/api-zod/src/generated/types/frequencyData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartBus AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FrequencyData {
  routeId: string;
  dayType: string;
  /** Buses per hour in morning (6-10am) */
  morning: number;
  /** Buses per hour in afternoon (10am-5pm) */
  afternoon: number;
  /** Buses per hour in evening (5-9pm) */
  evening: number;
  /** Buses per hour at night (9pm-midnight) */
  night: number;
}