# lib/api-zod/src/generated/types/liveBus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartBus AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LiveBusCrowdLevel } from "./liveBusCrowdLevel";

export interface LiveBus {
  id: string;
  routeId: string;
  routeNumber: string;
  routeName: string;
  routeColor: string;
  lat: number;
  lng: number;
  speed: number;
  heading: number;
  nextStop: string;
  nextStopId: string;
  distanceToNextStop: number;
  crowdLevel: LiveBusCrowdLevel;
  isLastBus: boolean;
  lastUpdated: string;
}